chore: add static anaysis

+++ src/BookStore/Application/Command/DiscountBookCommandHandler.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\BookStore\Application\Command;
 
+use App\BookStore\Domain\Exception\MissingBookException;
 use App\BookStore\Domain\Repository\BookRepositoryInterface;
 use App\Shared\Application\Command\CommandHandlerInterface;
 
@@ -16,6 +17,10 @@ final class DiscountBookCommandHandler implements CommandHandlerInterface
     public function __invoke(DiscountBookCommand $command): void
     {
         $book = $this->bookRepository->ofId($command->id);
+        if (null === $book) {
+            throw new MissingBookException($command->id);
+        }
+
         $book->applyDiscount($command->discount);
 
         $this->bookRepository->remove($book);
